# [Std] [lang] Add ti.rsqrt() and ti.Vector.norm_inv()