@archibate 2020-06-20
Running the following test code:
```py
import taichi as ti
import numpy as np
import time
ti.init(ti.opengl)
use_rsqrt = False

N, M = 2048, 256

x = ti.Vector(3, ti.f32, (N, M))

@ti.kernel
def func():
    for i in range(N):
        for j in range(M):
            if ti.static(use_rsqrt):
                x[i, j] = x[i, j] * x[i, j].norm_inv()
            else:
                x[i, j] = x[i, j] / x[i, j].norm()


x.from_numpy(np.random.random((N, M, 3)))
func()
func()
func()

print('starting...')
times = 0.0
for i in range(1000):
    a = time.time()
    func()
    if ti.cfg.arch == ti.cuda:
        ti.sync()
    elif ti.cfg.arch == ti.opengl:
        x.to_numpy()
    times += time.time() - a

print(times, 'ms')
```

obtains:
CPU: `0.88ms -> 0.71ms`.
CUDA: `1.58ms -> 1.57ms`.
Mesa OpenGL: `1.97ms -> 1.97ms`.
NVIDIA OpenGL: `3.01ms -> 3.01ms`.

Well... not sure if it's memory copy overhead or compiler optimization make GPU boost invisible...