@yuanming-hu 2020-06-20
> Well... not sure if it's memory copy overhead or compiler optimization make GPU boost invisible...

Your kernel is memory-bound. To make it compute-bound, do something like

```python
@ti.kernel
def func():
    for i in range(N):
        for j in range(M):
            if ti.static(use_rsqrt):
                x[i, j] = x[i, j] * (x[i, j].norm_inv() + x[i, j].norm_inv(eps=1e-1)+ x[i, j].norm_inv(eps=1e-2) + x[i, j].norm_inv(eps=1e-3)+ x[i, j].norm_inv(eps=1e-4))
            else:
                x[i, j] = x[i, j] * (1 / x[i, j].norm() + 1 / x[i, j].norm(eps=1e-1) + 1 / x[i, j].norm(eps=1e-2) + 1 / x[i, j].norm(eps=1e-3))
```

Also note that some backend compilers will do rsqrt for `1 / sqrt` automatically under `fastmath`.