@archibate 2020-06-20
Thank for the information! It's `2.0ms -> 1.6ms` on CPU now.