@archibate 2020-06-21
Great! Let's merge this for now and add rsqrt autodiff support iapr.